Add can_publish table column:
The new "can_publish" table column (and the correlating methods,
"enable_publish" and "disable_publish") allow you to distinguish
whether an author can publish articles or not.

This is useful for a blog in which you want contributors that write
articles but an administrator to hit the final "publish" button.