Set the "user" variable from the before hook:
This makes the code much cleaner and simpler. The RBAC auth
plugin now doesn't need to have access to the session or to a
handler that reaches into the schema.

This also allows to have a "user" variable where there is no
authentication running (like the main page).

Those can later be also made configurable.