Add a global variable to track if the site is static or not
When building a static site, the 'static' configuration setting is set
to true.

This will influence the templates (to hide the "dynamic" links, content,
and forms) and the URL generation (static sites need extensions).

Until we come up with a cleaner way to keep track of the 'static' setting,
it will be stored in a package variable when the application starts.

The package variable is set at compile time (after the Dancer
configuration has been loaded), which makes it available as early as
possible to the other modules.

Therefore, to detect if the app is running in 'static' mode, use:
- [% IF settings.static %]   in the templates
- if( $PearlBee::is_static ) in the Perl code