Refactor the commenting system into a separate "engine"
Up until now, PearlBee had it's own commenting system builtin. So that
commenting system had a database table "comments", which was right there
among the other tables. It had a few endpoints such as /comments/add,
which were right there among other endpoints, such as the post viewing
endpoint. The same happened in the presentation layer, so the templates
also had everything together.

What we do in this commit is separate everything that is related to
commenting. (Well, that's not entirely true, since the database table,
for example, is still in the same place.)

We create a PearlBee::Comments class, which is loaded when the
application starts. That class will check the config for "comments", and
will try to load the engine defined there. At this point, the only valid
value for that key is 'builtin', which means we will load
P::Comments::Builtin and P::Comments::Builtin::Endpoints. The first one
is required for every new engine, and the second one is optional.

P::Comments::Builtin will be instantiated with any config parameters
defined in comments_engines.builtin in config.yml, and will return an
object. That object will be set for every template, so that when a
template wants to display the comment count for a given post, all it has
to do is:

[% INCLUDE $comments_engine.post_comment_count_tt %]

Note that the $ sign there is not a typo :)

The other class, P::Comments::Builtin::Endpoints, will simply define all
the endpoints that this engine requires, such as add_comment, and things
in the dashboard to manage the comments.

There was some renaming done:

lib/PearlBee/Dashboard/Comments.pm -> lib/PearlBee/Comments/Builtin/Dashboard.pm
views/comment_form.tt -> views/comments/builtin/comment_form.tt
views/list_comments.tt -> views/comments/builtin/list_comments.tt

So don't despair if you see a long patch in git log -p. A big portion of
it is just renaming.