Move captcha to a separate feature, controlled by config
We create a new config parameter called "captcha_enabled", which is
set to false by default.
Now, we only load Authen::Captcha if that feature is enabled. This way,
we can skip installing GD development files unless this is strictly
necessary.

The rationale is this: imagine a blog which will not be open to user
registrations, or comments (for example, they might use Disqus or
IntenseDebate, or even facebook comments). So, no captchas necessary.

Now imagine a designer forking the repository to change stuff in the
templates. They cannot load the application because it fails with a
cryptic message about missing GD headers. They will probably have
no idea what that means, and why should they? Let them install the least
number of dependencies possible, preferrably all with cpanm :)