Remove the map_posts utility function
The map_posts() utility in PearlBee::Helpers::Util takes a list of Post
objects and returns a list of hashes.

Most of the hash keys are simply the post attributes. The 'content' key,
which contained an excerpt of the post text is commented out.

If the first lines of a post are needed, they should be provided by
a method in PearlBee::Model::Schema::Result::Post.